@@ -8,17 +8,17 @@
     "dist"
   ],
   "scripts": {
-    "build": "rollup -c rollup/prod.config.js",
-    "dev": "rollup -c rollup/dev.config.js -w",
+    "build": "rollup -c rollup/prod.config.ts",
+    "dev": "rollup -c rollup/dev.config.ts -w",
     "test": "nyc --check-coverage --reporter=lcov --reporter=text cypress run-ct --config-file=tests/cypress/cypress.json",
     "test:smoke": "npm run build && BABEL_ENV=production npm test -- --browser=chrome",
     "test-runner": "cypress open-ct --config-file=tests/cypress/cypress.json",
     "docs:dev": "vuepress dev docs",
     "docs:build": "vuepress build docs",
-    "validate-docs": "gulp --gulpfile=scripts/validateReadme.gulp.js",
+    "validate-docs": "npx ts-node scripts/validateReadme.ts",
     "commit": "git-cz",
     "prepare": "husky install",
-    "prepack": "gulp --gulpfile=scripts/copyDocsReadmeToRoot.gulp.js"
+    "prepack": "npx ts-node scripts/copyDocsReadmeToRoot.ts"
   },
   "peerDependencies": {
     "react": ">=16.8.0",
@@ -31,9 +31,9 @@
     "@babel/preset-typescript": "^7.13.0",
     "@commitlint/cli": "^12.1.1",
     "@commitlint/config-conventional": "^12.1.1",
-    "@cypress/code-coverage": "^3.9.10",
+    "@cypress/code-coverage": "^3.9.11",
     "@cypress/react": "^5.10.0",
-    "@cypress/webpack-dev-server": "^1.4.0",
+    "@cypress/webpack-dev-server": "^1.6.0",
     "@rollup/plugin-commonjs": "^18.0.0",
     "@rollup/plugin-html": "^0.2.3",
     "@rollup/plugin-node-resolve": "^11.2.1",
@@ -42,6 +42,7 @@
     "@semantic-release/github": "^7.0.7",
     "@semantic-release/npm": "^7.1.1",
     "@semantic-release/release-notes-generator": "^9.0.1",
+    "@types/node": "^16.10.1",
     "@types/react": "^17.0.3",
     "@typescript-eslint/eslint-plugin": "^4.22.0",
     "@typescript-eslint/parser": "^4.22.0",
@@ -63,8 +64,6 @@
     "eslint-plugin-react": "^7.23.2",
     "eslint-plugin-react-hooks": "^3.0.0",
     "gh-pages": "^3.2.3",
-    "gulp": "^4.0.2",
-    "gulp-replace": "^1.1.3",
     "husky": "^5.1.3",
     "lint-staged": "^10.2.6",
     "nyc": "^15.1.0",
@@ -78,13 +77,13 @@
     "rollup-plugin-serve": "^1.1.0",
     "rollup-plugin-typescript2": "^0.30.0",
     "semantic-release": "^17.4.2",
-    "through2": "^4.0.2",
+    "ts-node": "^10.2.1",
     "tslib": "^2.2.0",
-    "typescript": "^4.2.4",
+    "typescript": "^4.4.3",
     "uuid": "^8.3.2",
     "vuepress": "^2.0.0-beta.16",
-    "webpack": "^5.41.1",
-    "webpack-dev-server": "^3.11.2"
+    "webpack": "^5.55.1",
+    "webpack-dev-server": "^4.3.0"
   },
   "author": "Ryan Geyer",
   "homepage": "https://react-hover-video-player.dev",
